feat(aapd-1195): move database to shared package
