+++ azure-pipelines-db-seed.yml
@@ -43,21 +43,21 @@ jobs:
             - template: steps/azure_get_secrets.yml@templates
               parameters:
                 secrets:
-                  - name: appeals-sql-server-connection-string-admin
-                    variable: SQL_CONNECTION_STRING_ADMIN
+                  - name: appeals-sql-server-connection-string-app
+                    variable: SQL_CONNECTION_STRING
             - template: steps/node_script.yml@templates
               parameters:
                 condition: ne('${{ parameters.environment }}', 'Prod')
                 environmentVariables:
-                  SQL_CONNECTION_STRING_ADMIN: $(SQL_CONNECTION_STRING_ADMIN)
+                  SQL_CONNECTION_STRING: $(SQL_CONNECTION_STRING)
                 script: npm run db:generate && npm run db:seed
-                workingDirectory: $(Build.Repository.LocalPath)/packages/appeals-service-api
+                workingDirectory: $(Build.Repository.LocalPath)/packages/database
             - template: steps/node_script.yml@templates
               parameters:
                 condition: eq('${{ parameters.environment }}', 'Prod')
                 environmentVariables:
-                  SQL_CONNECTION_STRING_ADMIN: $(SQL_CONNECTION_STRING_ADMIN)
+                  SQL_CONNECTION_STRING: $(SQL_CONNECTION_STRING)
                 script: npm run db:seed:prod
-                workingDirectory: $(Build.Repository.LocalPath)/packages/appeals-service-api
+                workingDirectory: $(Build.Repository.LocalPath)/packages/database
     workspace:
       clean: all
