+++ package.json
@@ -20,11 +20,11 @@
 		"docker:comment": "docker compose --profile comment -f docker-compose.yml -f docker-compose.slim.yml up",
 		"commit": "cz",
 		"commitlint": "commitlint --from=$(git log origin/main..$(git rev-parse HEAD) --pretty=format:\"%H\" | tail -n 1)",
-		"db:migrate:dev": "npm run db:migrate:dev --workspace appeals-service-api",
-		"db:migrate:prod": "npm run db:migrate:prod --workspace appeals-service-api",
-		"db:generate": "npm run db:generate --workspace appeals-service-api",
-		"db:seed": "npm run db:seed --workspace appeals-service-api",
-		"db:seed:prod": "npm run db:seed:prod --workspace appeals-service-api",
+		"db:migrate:dev": "npm run migrate:dev --workspace @pins/database",
+		"db:migrate:prod": "npm run migrate:prod --workspace @pins/database",
+		"db:generate": "npm run generate --workspace @pins/database",
+		"db:seed": "npm run seed --workspace @pins/database",
+		"db:seed:prod": "npm run seed:prod --workspace @pins/database",
 		"format": "prettier --write \"**/*.{js,ts,json,md}\"",
 		"lint": "run-s --continue-on-error --print-label lint:appeals-api lint:business-rules lint:common lint:documents-api lint:pdf-api lint:web lint:integration-functions",
 		"lint:appeals-api": "cd packages/appeals-service-api; eslint ./",
