+++ packages/appeals-service-api/Dockerfile
@@ -15,11 +15,13 @@ COPY package.json /opt/
 COPY package-lock.json /opt/
 COPY packages/common /opt/packages/common
 COPY packages/business-rules /opt/packages/business-rules
+COPY packages/database /opt/packages/database
 
 # install from root
 WORKDIR /opt
 RUN npm ci --omit=dev
 # generate prisma client
+WORKDIR /opt/packages/appeals-service-api
 RUN npm run db:generate
 
 # Stage 2/2: App Run
