--- packages/appeals-service-api/src/db/seed/seed-prod.js
@@ -1,17 +1,0 @@
-const { createPrismaClient } = require('../db-client');
-const { seedStaticData } = require('./data-static');
-
-async function run() {
-	const dbClient = createPrismaClient();
-
-	try {
-		await seedStaticData(dbClient);
-	} catch (error) {
-		console.error(error);
-		throw error;
-	} finally {
-		await dbClient.$disconnect();
-	}
-}
-
-run();
