+++ packages/appeals-service-api/src/routes/v2/users/service.js
@@ -1,6 +1,6 @@
 const { AppealUserRepository } = require('#repositories/sql/appeal-user-repository');
 const ApiError = require('#errors/apiError');
-const { APPEAL_USER_ROLES_ARRAY } = require('../../../db/seed/data-static');
+const { APPEAL_USER_ROLES_ARRAY } = require('@pins/database/src/seed/data-static');
 
 const appealUserRepository = new AppealUserRepository();
