+++ packages/appeals-service-api/src/routes/v2/users/users.spec.js
@@ -4,7 +4,7 @@ const supertest = require('supertest');
 
 const app = require('../../../app');
 const { createPrismaClient } = require('../../../db/db-client');
-const { seedStaticData } = require('../../../db/seed/data-static');
+const { seedStaticData } = require('@pins/database/src/seed/data-static');
 
 const { isFeatureActive } = require('../../../configuration/featureFlag');
