+++ packages/database/src/seed/seed-dev.js
@@ -1,9 +1,14 @@
-const { createPrismaClient } = require('../db-client');
+const createPrismaClient = require('../create-client');
 const { seedStaticData } = require('./data-static');
 const { seedDev } = require('./data-dev');
+const config = require('../configuration/config');
 
 async function run() {
-	const dbClient = createPrismaClient();
+	const prismaConfig = {
+		datasourceUrl: config.db.sql.connectionString
+	};
+
+	const dbClient = createPrismaClient(prismaConfig);
 
 	try {
 		await seedStaticData(dbClient);
